requirements.txt
asgiref==3.2.10
Django==3.0.8
django-restframework==0.0.1
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
